+++ main/src/components/ActionBar/SourceVerificationIcon/SourceVerificationIcon.jsx
@@ -1,12 +1,11 @@
-import { CircularProgress, Tooltip } from "@mui/material";
+import { CircularProgress, Tooltip, IconButton } from "@mui/material";
 import { Component, useState } from "react";
 import QuestionMarkIcon from "@mui/icons-material/QuestionMark";
-import { Button } from "react-admin";
 import PropTypes from "prop-types";
 import GppGoodIcon from '@mui/icons-material/GppGood';
 import GppBadIcon from '@mui/icons-material/GppBad';
 import GppMaybeIcon from '@mui/icons-material/GppMaybe';
-import myVerify from '../../../../src/vendor/verify';
+import { coreVerify } from '../../../vendor/vcCore';
 import comunicaEngineWrapper from '../../../comunicaEngineWrapper/comunicaEngineWrapper';
 
 const VERIFICATION_STATES = {
@@ -43,7 +42,7 @@ function SourceVerificationIcon({ context, source, proxyUrl }) {
     try {
       const response = await fetchFunction(source);
       const verifiableCredential = await response.json();
-      const { validationResult, verificationResult } = await myVerify(verifiableCredential);
+      const { validationResult, verificationResult } = await coreVerify(verifiableCredential);
       if (validationResult.valid) {
         if (verificationResult.verified) {
           return VERIFICATION_STATES.VERIFIED;
@@ -100,9 +99,9 @@ function SourceVerificationIcon({ context, source, proxyUrl }) {
   } else {
     return (
       <Tooltip title="Verify source">
-        <Button onClick={verify}>
-          <QuestionMarkIcon size="small" />
-        </Button>
+        <IconButton sx={{ color: "#1976D2", padding: "0", marginLeft: "0" }} onClick={verify}>
+          <QuestionMarkIcon fontSize="small" />
+        </IconButton>
       </Tooltip>
     );
   }