+++ main/src/dataProvider/SparqlDataProvider.js
@@ -1,5 +1,4 @@
 import { ProxyHandlerStatic } from "@comunica/actor-http-proxy";
-import { HttpError } from "react-admin";
 import { Generator, Parser } from "sparqljs";
 import NotImplementedError from "../NotImplementedError";
 import { Term } from "sparqljs";
@@ -27,8 +26,26 @@ const onConfigChanged = (newConfig) => {
 
 configManager.on('configChanged', onConfigChanged);
 
+// simple cache to save time while scrolling through pages of a list
+// results = the result of executeQuery, totalItems
+const listCache = {
+  hash: "",
+  results: {}
+};
+
+// LOG let getListCounter = 0;
+// LOG let getVariableOptionsCounter = 0;
+
 export default {
   getList: async function getList(resource, { pagination, sort, filter, meta }) {
+
+    // LOG console.log(`--- getList #${++getListCounter}`);
+    // LOG console.log(`resource: ${ resource }`);
+    // LOG console.log(`pagination: ${JSON.stringify(pagination, null, 2)}`);
+    // LOG console.log(`sort: ${ JSON.stringify(sort, null, 2) }`);
+    // LOG console.log(`filter: ${JSON.stringify(filter, null, 2)}`);
+    // LOG console.log(`meta: ${JSON.stringify(meta, null, 2)}`);
+    
     // make a working copy of the query object found in the configuration, to prevent changing the configuration
     // this copy is extended here
     // rendering should occur based on this working copy
@@ -44,11 +61,23 @@ export default {
       query.comunicaContext.sources = [...new Set([...query.comunicaContext.sources, ...additionalSources])];
     }
 
-    if (meta && meta.variables) {
-      query.variableValues = meta.variables;
+    if (meta && meta.variableValues) {
+      query.variableValues = meta.variableValues;
     }
 
-    let results = await executeQuery(query);
+    let results;
+    const hash = JSON.stringify({ resource, variableValues: query.variableValues });
+    // LOG console.log(`hash: ${hash}`);
+    if (hash == listCache.hash) {
+      // LOG console.log(`reusing listCache.results: ${JSON.stringify(listCache.results, null, 2)}`);
+      results = listCache.results;
+    } else {
+      results = await executeQuery(query);
+      listCache.hash = hash;
+      listCache.results = results;
+      // LOG console.log(`new listCache.results: ${JSON.stringify(listCache.results, null, 2)}`);
+    }
+
     let totalItems = results.length;
     results = results.slice(offset, offset + limit);
 
@@ -91,9 +120,8 @@ export default {
   deleteMany: async function deleteMany() {
     throw new NotImplementedError();
   },
-  indirectVariables: async function indirectVariables(query) {
-    return await getIndirectVariables(query);
-  },
+  getVariableOptions,
+  clearListCache
 };
 
 /**
@@ -135,7 +163,7 @@ async function buildQueryText(query) {
     const generator = new Generator();
     return generator.stringify(parsedQuery);
   } catch (error) {
-    throw new HttpError(error.message, 500);
+    throw new Error(error.message);
   }
 }
 
@@ -143,11 +171,11 @@ async function buildQueryText(query) {
  * Replace the variable placeholders in a query's raw text by the specified value.
  * 
  * @param {string} rawText - the raw text of a query.
- * @param {object} variables - an object containing the variable names and specified values (as strings).
- * @returns {string} the resulting raw text of the query after replacing the variables.
+ * @param {object} variableValues - an object containing the variable names and specified values (as strings).
+ * @returns {string} the resulting raw text of the query after replacing the variableValues.
  */
-function replaceVariables(rawText, variables) {
-  for (const [variableName, variableValue] of Object.entries(variables)) {
+function replaceVariables(rawText, variableValues) {
+  for (const [variableName, variableValue] of Object.entries(variableValues)) {
     // do not surround with double quotes here; add double quotes in the input if needed!
     rawText = rawText.replaceAll("$" + variableName, variableValue);
   }
@@ -214,13 +242,25 @@ async function executeQuery(query) {
       }
       results.push(newResult);
     };
+    const callbackBoolean = (b) => {
+      const answer = b ? query?.askQuery?.trueText || "Yes" : query?.askQuery?.falseText || "No";
+      const newResult = {
+        Answer: {
+          termType: "Literal",
+          value: answer,
+          language: ""
+        },
+        id: index++
+      }
+      results.push(newResult);
+    };
     await comunicaEngineWrapper.query(query.queryText,
        // WEIRD: we need to make a copy of the context here (a shallow copy is fine); concurrent calls ???
       { ...query.comunicaContext },
-      { "variables": callbackVariables, "bindings": callbackBindings, "quads": callbackQuads });
+      { "variables": callbackVariables, "bindings": callbackBindings, "quads": callbackQuads, "boolean": callbackBoolean });
     return results;
   } catch (error) {
-    throw new HttpError(error.message, 500);
+    throw new Error(`Error executing SPARQL query "${query.rawText || query.queryString}": ${error.message}`);
   }
 }
 
@@ -295,26 +335,27 @@ function handleComunicaContextCreation(query) {
   }
 }
 
-async function getIndirectVariables(query) {
+async function getVariableOptions(query) {
 
+  // LOG console.log(`--- getVariableOptions #${++getVariableOptionsCounter}`);
 
-  // This chunk of code is duplicated in order for the indirect queries having sources from a source index to work correctly.
+  // BEGIN duplicated chunk of code (duplicated in order for templated queries with indirect queries having sources from a source index to work correctly)
   handleComunicaContextCreation(query);
 
   if (query.sourcesIndex) {
     const additionalSources = await getSourcesFromSourcesIndex(query.sourcesIndex, query.comunicaContext.useProxy);
     query.comunicaContext.sources = [...new Set([...query.comunicaContext.sources, ...additionalSources])];
   }
-  // 
+  // END duplicated chunk of code
 
 
-  let variables;
+  let variableOptions;
   let queryStringList = [];
 
   if (query.variables) {
-    variables = query.variables
+    variableOptions = query.variables;
   } else {
-    variables = {}
+    variableOptions = {};
   }
 
   if (query.indirectVariables) {
@@ -351,8 +392,8 @@ async function getIndirectVariables(query) {
           // see https://comunica.dev/docs/query/advanced/bindings/
           for (const [variable, term] of bindings) {
             const name = variable.value;
-            if (!variables[name]) {
-              variables[name] = [];
+            if (!variableOptions[name]) {
+              variableOptions[name] = [];
             }
             let variableValue;
             switch (term.termType) {
@@ -387,8 +428,8 @@ async function getIndirectVariables(query) {
               default:
                 break;
             }
-            if (variableValue && !variables[name].includes(variableValue)) {
-              variables[name].push(variableValue);
+            if (variableValue && !variableOptions[name].includes(variableValue)) {
+              variableOptions[name].push(variableValue);
             }
           }
         });
@@ -398,11 +439,20 @@ async function getIndirectVariables(query) {
     }
   }
   catch (error) {
-    throw new Error(`Error adding indirect variables: ${error.message}`);
+    throw new Error(`Error getting variable options: ${error.message}`);
   }
 
-  if (variables == {}) {
-    throw new Error(`The variables are empty`);
+  if (variableOptions == {}) {
+    throw new Error(`The variable options are empty`);
   }
-  return variables;
+  return variableOptions;
 }
+
+/**
+ * Clears the list cache
+ */
+function clearListCache() {
+  // LOG console.log('Clearing listCache');
+  listCache.hash = "";
+}
+